+++ app/Http/Controllers/DashboardPostController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use \Cviebrock\EloquentSluggable\Services\SlugService;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
 
 class DashboardPostController extends Controller
 {
@@ -16,13 +17,15 @@ class DashboardPostController extends Controller
      */
     public function index()
     {
-        $this->authorize('admin');
-        
-        return view('dashboard.post.index', [
-            'posts' => Post::where('user_id', auth()->user()->id)
-            ->orderByDesc('created_at')
-            ->get()
-        ]);
+        if (Gate::denies('user')) {
+            return view('dashboard.post.index', [
+                'posts' => Post::where('user_id', auth()->user()->id)
+                ->orderByDesc('created_at')
+                ->get(),
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -30,9 +33,13 @@ class DashboardPostController extends Controller
      */
     public function create()
     {
-        return view('dashboard.post.create', [
-            'categories' => Category::all(),
-        ]);
+        if (Gate::denies('user')) {
+            return view('dashboard.post.create', [
+                'categories' => Category::all(),
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -40,26 +47,32 @@ class DashboardPostController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('user')) {
+            $validatedData = $request->validate([
+                'title' => 'required|max:255',
+                'slug' => 'required|unique:posts',
+                'category_id' => 'required',
+                'image' => 'image|file|max:1500',
+                'body' => 'required',
+            ], [
+                'required' => 'The :attribute field is required.',
+                'max' => 'The :attribute may not be greater than :max kilobytes.',
+            ]);
 
-        $validatedData = $request->validate([
-            'title' => 'required|max:255',
-            'slug' => 'required|unique:posts',
-            'category_id' => 'required',
-            'image' => 'image|file|max:1500',
-            'body' => 'required',
-        ]);
+            if($request->file('image'))
+            {
+                $validatedData['image'] = $request->file('image')->store('post-images');
+            }
 
-        if($request->file('image'))
-        {
-            $validatedData['image'] = $request->file('image')->store('post-images');
+            $validatedData['user_id'] = auth()->user()->id;
+            $validatedData['excerpt'] = Str::limit(strip_tags($request->body), 200);
+
+            Post::create($validatedData);
+            
+            return redirect('/dashboard/posts')->with('success', 'Post has been created!');
+        } else {
+            abort(403);
         }
-
-        $validatedData['user_id'] = auth()->user()->id;
-        $validatedData['excerpt'] = Str::limit(strip_tags($request->body), 200);
-
-        Post::create($validatedData);
-        
-        return redirect('/dashboard/posts')->with('success', 'Post has been created!');
     }
 
     /**
@@ -67,9 +80,17 @@ class DashboardPostController extends Controller
      */
     public function show(Post $post)
     {
-        return view('dashboard.post.show', [
-            'post' => $post
-        ]);
+        if (auth()->user()->id !== $post->user_id) {
+            abort(403);
+        }
+
+        if (Gate::denies('user')) {
+            return view('dashboard.post.show', [
+                'post' => $post
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -77,10 +98,18 @@ class DashboardPostController extends Controller
      */
     public function edit(Post $post)
     {
-        return view('dashboard.post.edit', [
-            'post' => $post,
-            'categories' => Category::all(),
-        ]);
+        if (auth()->user()->id !== $post->user_id) {
+            abort(403);
+        }
+
+        if (Gate::denies('user')) {
+            return view('dashboard.post.edit', [
+                'post' => $post,
+                'categories' => Category::all(),
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -88,36 +117,40 @@ class DashboardPostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        $rules = [
-            'title' => 'required|max:255',
-            'category_id' => 'required',
-            'image' => 'image|mimes:jpeg,png,jpg|max:1500',
-            'body' => 'required',
-        ];
+        if (Gate::denies('user')) {
+            $rules = [
+                'title' => 'required|max:255',
+                'category_id' => 'required',
+                'image' => 'image|mimes:jpeg,png,jpg|max:1500',
+                'body' => 'required',
+            ];
 
-        if($request->slug != $post->slug)
-        {
-            $rules ['slug'] = 'required|unique:posts';
-        }
-
-        $validatedData = $request->validate($rules);
-
-        $validatedData['user_id'] = auth()->user()->id;
-        $validatedData['excerpt'] = Str::limit(strip_tags($request->body), 200);
-
-        if($request->file('image'))
-        {
-            if($request->oldImage) {
-                Storage::delete($request->oldImage);
+            if($request->slug != $post->slug)
+            {
+                $rules ['slug'] = 'required|unique:posts';
             }
 
-            $validatedData['image'] = $request->file('image')->store('post-images');
+            $validatedData = $request->validate($rules);
+
+            $validatedData['user_id'] = auth()->user()->id;
+            $validatedData['excerpt'] = Str::limit(strip_tags($request->body), 200);
+
+            if($request->file('image'))
+            {
+                if($request->oldImage) {
+                    Storage::delete($request->oldImage);
+                }
+
+                $validatedData['image'] = $request->file('image')->store('post-images');
+            }
+
+            Post::where('id', $post->id)
+                ->update($validatedData);
+
+            return redirect('/dashboard/posts')->with('success', 'Post has been edited!');
+        } else {
+            abort(403);
         }
-
-        Post::where('id', $post->id)
-            ->update($validatedData);
-
-        return redirect('/dashboard/posts')->with('success', 'Post has been edited!');
 
     }
 
@@ -126,15 +159,18 @@ class DashboardPostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Gate::denies('user')) {
+            if($post->image)
+            {
+                Storage::delete($post->image);
+            }
 
-        if($post->image)
-        {
-            Storage::delete($post->image);
+            Post::destroy($post->id);
+
+            return redirect('/dashboard/posts')->with('success', 'Post has been deleted!');
+        } else {
+            abort(403);
         }
-
-        Post::destroy($post->id);
-
-        return redirect('/dashboard/posts')->with('success', 'Post has been deleted!');
     }
 
     public function checkSlug(Request $request)
